Prepare for removing dependency to Bootstrap
