+++ src/components/_components-theming-mixins.scss
@@ -2,7 +2,6 @@
 @import "./group/group-theming-mixins";
 @import "./image/image-theming-mixins";
 @import "./loading/loading-theming-mixins";
-@import "./tab/tab-theming-mixins";
 @import "./table/table-theming-mixins";
 @import "./textblock/textblock-theming-mixins";
 @import "./chronometer/chronometer-theming-mixins";
@@ -11,6 +10,7 @@
 @import "./renders/bootstrap/form-field/form-field-theming-mixins";
 @import "./renders/bootstrap/button/button-theming-mixins";
 @import "./renders/bootstrap/card/card-theming-mixins";
-@import "./renders/bootstrap/tab-caption/tab-caption-theming-mixins";
+@import "./tab/tab-theming-mixins";
+@import "./tab-caption/tab-caption-theming-mixins";
 @import "./navbar/navbar-theming-mixins";
 @import "./action-sheet/action-sheet-theming-mixins";
