--- src/components/renders/bootstrap/tab-caption/tab-caption-render.scss
@@ -1,5 +1,0 @@
-gx-tab-caption {
-  .nav-link {
-    display: flex;
-  }
-}
