--- src/components/renders/bootstrap/tab/tab-render.scss
@@ -1,13 +1,0 @@
-gx-tab {
-  --elevation: 0;
-  --tab-strip-elevation: 0;
-
-  & > div {
-    @include elevation();
-
-    & > .nav-tabs {
-      --elevation: var(--tab-strip-elevation);
-      @include elevation();
-    }
-  }
-}
