+++ src/components/tab-caption/_tab-caption-theming-mixins.scss
