+++ src/components/tab-caption/tab-caption-render.e2e.ts
